+++ ios/Classes/FlutterRTCVideoPlatformViewController.m
@@ -1,0 +1,1 @@
+../../common/darwin/Classes/FlutterRTCVideoPlatformViewController.m